server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
azure-ai-documentintelligence>=1.0.0b4
azure-core>=1.30.0
openai>=1.0.0
deepgram-sdk>=3.0.0
pyaudio>=0.2.11
pillow>=10.0.0
pymupdf>=1.23.0
